# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API da loja de chapéus Hat Store.",
        "title": "Hat Store API",
        "contact": {},
        "version": "1.0"
    },
    "host": "hatstore-prd.fly.dev",
    "basePath": "/api/",
    "paths": {
...
        "/cart/add": {
            "post": {
                "description": "Adiciona um chapéu ao carrinho do usuário",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "Adiciona um chapéu ao carrinho",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "Item do carrinho",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Hat"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/cart/update": {
            "put": {
                "description": "Atualiza a quantidade de um item no carrinho",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "Atualiza o carrinho",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "Item do carrinho",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Hat"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/checkout": {
            "post": {
                "description": "Finaliza o pedido do usuário",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "checkout"
                ],
                "summary": "Finaliza o pedido",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "Dados do pedido",
                        "name": "pedido",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/estoque": {
            "get": {
                "description": "Retorna o estoque atual de cada chapéu",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "estoque"
                ],
                "summary": "Lista o estoque de chapéus",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID de desenvolvimento",
                        "name": "X-Dev-UUID",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.Hat"
                            }
                        }
                    }
                }
            }
        },
        "/hats": {
            "get": {
                "description": "Retorna a lista de chapéus disponíveis",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hats"
                ],
                "summary": "Lista todos os chapéus",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID de desenvolvimento",
                        "name": "X-Dev-UUID",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Hat"
                            }
                        }
                    }
                }
            }
        },
        "/pedido": {
            "post": {
                "description": "Registra os dados do pedido em memória",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pedido"
                ],
                "summary": "Registra um pedido",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID de desenvolvimento",
                        "name": "X-Dev-UUID",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "JWT token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "Dados do pedido",
                        "name": "pedido",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Pedido"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Pedido"
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Acesso não autorizado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Produto sem estoque suficiente",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.Hat": {
            "type": "object",
            "properties": {
                "categoria": {
                    "description": "nacional, importado, crescer",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "quantidade": {
                    "type": "integer"
                }
            }
        },
...
        "models.Hat": {
            "type": "object",
            "properties": {
                "categoria": {
                    "description": "nacional, importado, crescer",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "quantidade": {
                    "type": "integer"
                }
            }
        },
...
        "models.Pedido": {
            "type": "object",
            "properties": {
                "bairro": {
                    "type": "string"
                },
                "cep": {
                    "type": "string"
                },
                "cidade": {
                    "type": "string"
                },
                "cpf": {
                    "type": "string"
                },
                "cupom": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "endereco": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "itens": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Hat"
                    }
                },
                "nome": {
                    "type": "string"
                },
                "numero": {
                    "type": "string"
                },
                "pagamento": {
                    "description": "\"pix\" ou \"boleto\"",
                    "type": "string"
                },
                "telefone": {
                    "type": "string"
                },
                "total": {
                    "type": "number"
                },
                "uf": {
                    "type": "string"
                }
            }
        },
...
    }
}